renamed localeTime to time
for the sake of elegant looking components folder
