+++ components/time/index.ts
